Moving Collection extractors to their own folder
