+++ src/als-tools.infrastructure/Extractors/Collections/LiveProjectsCollectionExtractor.cs
@@ -1,6 +1,6 @@
 using AlsTools.Core.Entities;
 
-namespace AlsTools.Infrastructure.Extractors;
+namespace AlsTools.Infrastructure.Extractors.Collections;
 
 public interface ILiveProjectsCollectionExtractor : ICollectionExtractor<LiveProject>
 {
