+++ src/als-tools.infrastructure/Extractors/Collections/LocatorsCollectionExtractor.cs
@@ -1,6 +1,6 @@
 using AlsTools.Core.ValueObjects;
 
-namespace AlsTools.Infrastructure.Extractors;
+namespace AlsTools.Infrastructure.Extractors.Collections;
 
 public interface ILocatorsCollectionExtractor : ICollectionExtractor<Locator>
 {
