+++ src/als-tools.infrastructure/Extractors/Collections/ScenesCollectionExtractor.cs
@@ -1,6 +1,6 @@
 using AlsTools.Core.ValueObjects;
 
-namespace AlsTools.Infrastructure.Extractors;
+namespace AlsTools.Infrastructure.Extractors.Collections;
 
 public interface IScenesCollectionExtractor : ICollectionExtractor<Scene>
 {
