+++ src/als-tools.infrastructure/Extractors/Collections/TracksCollectionExtractor.cs
@@ -2,7 +2,7 @@ using AlsTools.Core.Factories;
 using AlsTools.Core.ValueObjects;
 using AlsTools.Core.ValueObjects.Tracks;
 
-namespace AlsTools.Infrastructure.Extractors;
+namespace AlsTools.Infrastructure.Extractors.Collections;
 
 public interface ITracksCollectionExtractor : ICollectionExtractor<ITrack>
 {
